Refactor API URL to use dynamically obtained outbound IP address
Instead of using a hardcoded IP address in the  variable, this commit modifies the code to dynamically obtain the outbound IP address using the  function. The  variable is updated to construct the URL with the dynamically obtained IP address. This change allows for greater flexibility and adaptability when the machine's IP address changes.